Update package structure
Move the scripts from the `rgain3.scripts` module into `rgain3` and the
rest of the code into the `rgain3.lib` module.

This means that the scripts `replaygain` and `collectiongain` can be
invoked using `python -m rgain3.replaygain` and `python -m
rgain3.collectiongain`.

Also converted the scripts into console entrypoints.
