+++ rgain3/lib/util.py
